docs(breadcrumb): update breadcrumb docs and structure
