+++ .storybook/main.js
@@ -4,7 +4,8 @@ module.exports = {
     '../projects/canopy/src/lib/accordion/accordion.stories.ts',
     '../projects/canopy/src/lib/alert/alert.stories.ts',
     '../projects/canopy/src/lib/brand-icon/brand-icons.stories.ts',
-    '../projects/canopy/src/lib/breadcrumb/breadcrumb.stories.ts',
+    '../projects/canopy/src/lib/breadcrumb/docs/guide.stories.mdx',
+    '../projects/canopy/src/lib/breadcrumb/docs/breadcrumb.stories.ts',
     '../projects/canopy/src/lib/button/docs/guide.stories.mdx',
     '../projects/canopy/src/lib/button/docs/button.stories.ts',
     '../projects/canopy/src/lib/button/docs/button-group.stories.ts',
